refactor(chat): extract prompt turn runner
